+++ src/main/java/org/owasp/securityshepherd/repository/ClassRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClassRepository extends CrudRepository<ClassEntity, Long> {
+public interface ClassRepository extends CrudRepository<ClassEntity, Integer> {
 
 	@Query("select count(1) from class where name = :name")
 	public boolean existsByName(String name);
