+++ src/main/java/org/owasp/securityshepherd/repository/SubmissionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubmissionRepository extends CrudRepository<Submission, Long> {
+public interface SubmissionRepository extends CrudRepository<Submission, Integer> {
 
 }
